requirements.txt
# Core Web Framework
flask==3.1.0
flask-socketio==5.3.6
eventlet==0.35.2

# Browser Automation (Legacy - Will be phased out)
selenium==4.33.0
webdriver-manager==4.0.2

# AI APIs
google-generativeai==0.8.3
openai==1.58.1

# Security
cryptography==43.0.3
python-dotenv==1.0.1

# Utilities
requests==2.32.4
aiohttp==3.11.11
asyncio==3.4.3

# Logging & Monitoring
colorlog==6.9.0

# Data Processing
pandas==2.2.3
numpy==2.2.1

# Testing (Development)
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0